+++ database/migrations/2024_03_25_133341_create_sale_items_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->unsignedBigInteger(SaleItemTable::SALE_ID);
             $table->unsignedBigInteger(SaleItemTable::PRODUCT_ID);
             $table->double(SaleItemTable::MRP);
-            $table->unsignedBigInteger(SaleItemTable::QUANTITY);
+            $table->unsignedInteger(SaleItemTable::QUANTITY);
             $table->unsignedBigInteger(SaleItemTable::FINANCE_YEAR);
             $table->timestamps();
